Extract images from bags/stereo/requirements.txt
cv_bridge==1.16.2
numpy==1.24.1
numpy==1.23.1
numpy==1.17.4
opencv_python==4.6.0.66
opencv_python_headless==4.7.0.72
piexif==1.1.3
progressbar33==2.4
PyYAML==6.0.1
rosbag==1.16.0
